dispatch refactor
Create a new hook `modal-action` that returns functions that call the
dispatch function to interact with the store, This allows us no longer
pass in too many props into the `Modal` component, it itself can call
the hooks based on the action passed in.

Remove multiple booleans for the modal render. Instead provide a modal
infomration object, which when exists, renders the modal and places the
appropriate action.

Removed all the document funcitons within `App.tsx`, `Menu.tsx`, and
`TopBar.tsx` and place them within the hook in `document-action`. So
instead of calling the dispatch and store into every component, only the
hook has to be called and the needed functionality can be extracted.

Updated Cypress test to ensure app functionality remained consistent.